requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
lightgbm>=4.0.0
ccxt>=4.3.0
pyarrow>=14.0.0
tqdm>=4.66.0
streamlit>=1.36.0
plotly>=5.24.0
yfinance>=0.2.40
google-generativeai>=0.7.2
requests>=2.32.3